+++ Sources/PaymentButtons/PaymentButton.swift
@@ -1,5 +1,9 @@
 import UIKit
+#if canImport(CorePayments)
+import CorePayments
+#endif
 
+// swiftlint:disable type_body_length
 /// Handles functionality shared across payment buttons
 public class PaymentButton: UIButton {
 
@@ -20,6 +24,12 @@ public class PaymentButton: UIButton {
     static let bundle = Bundle(for: PaymentButton.self)
     #endif
 
+    // Use an empty config and default to live environment for button analytics
+    private let analyticsService = AnalyticsService(
+        coreConfig: .init(clientID: "N/A", environment: .live),
+        orderID: "N/A"
+    )
+
     // MARK: - Init
 
     required init(
@@ -36,7 +46,9 @@ public class PaymentButton: UIButton {
         self.size = size
         self.insets = insets
         self.label = label
+        self.analyticsService.sendEvent("payment-button:initialized", buttonType: fundingSource.rawValue)
         super.init(frame: .zero)
+        self.addTarget(self, action: #selector(onTap), for: .touchUpInside)
     }
 
     // MARK: - Views
@@ -155,6 +167,12 @@ public class PaymentButton: UIButton {
             }
             return false
         }
+    }
+
+    // MARK: - Private
+
+    @objc private func onTap() {
+        analyticsService.sendEvent("payment-button:tapped", buttonType: fundingSource.rawValue)
     }
 
     // MARK: - Configuration